various: change EXIF metadata into AVFrameSideData
This patch centralizes much of the EXIF parsing and handling code for
libavcodec, and delegates its own AVFrameSideData type to containing
the buffer that holds EXIF metadata. This patch also adds exposes the
exif parsing routing so it can be called by ffprobe, and updates the
corresponding FATE tests to read the keys from the side data instead of
from the main frame metadata.

This commit also removes an avpriv_ function in exif.h, exposing the
parsing functionality as a public API in the exported libavcodec/exif.h
header file. As such, this commit requires an ABI break and can only be
applied during a major version bump.

Signed-off-by: Leo Izen <leo.izen@gmail.com>