avcodec/exif: add EXIF parser
This commit adds a structure to contain parsed EXIF
metadata, as well as code to read and write that struct
from/to binary EXIF buffers. Some internal functions
have been moved to exif_internal.h. Code to read from
this new struct and write to an AVDictionary **dict has
been added as well in order to preserve interoperability
with existing callers. The only codec changes so far as of
this commit are to call these interop functions, but in future
commits there will be codec changes to use the new parsing
routines instead.

Signed-off-by: Leo Izen <leo.izen@gmail.com>